Refactor: separate page_screen package from page
Previously, not all classes in the `page` package were strictly related
to the implementor of `Page` in the GUI.

They are now moved into a separate `page_screen` package and have the
`GUI` prefix removed from their names.

The class which was previously called `GUIPage` is now more
appropriately titled `PageScreen` while the new `GUIPage` class is
strictly an implementor of the `Page` interface.