Lots of work: Pen tool and Storage
Writing a storage backend is hard when you don't have anything to store,
so I implemented a basic "pen" tool for drawing lines.

Added SQLite as a dependency to implement the SQLiteStorage class. At
this point, media objects can be saved/loaded and this is verified in
the TestSQLiteStorage unit test class.

Integrating SQLiteStorage into the rest of the program is not yet
complete.