Lots of changes: implement persistence + more
Get data persistence implemnted to the point where you can save/load
Media to files and this behaviour is usable from the GUI.

Also added a colour tool and added the settings GUI for the pen tool.
The pen tool now uses the colour configured with the colour tool.

Moved the main class to resolve an issue with loading JavaFX when try to
run the project from a jar file.