Fix bounds checking
This one was a little tricky. Initially, the implementation assumed that
the "translateX" and "translateY" for a Node always refers to the
position of the top left corner, but this is not actually the case and
the point at (translateX, translateY) can actually be anywhere within
the bounds of a node.

The bounds checking method in Media and the relevant SQL query in
SQLiteStorage are updated to address the above.