Lossless storage of tag strings
Previously, tag strings were just joined into a single string with
commas when stored in the SQLite db, but this could lead to a loss of
information if a tag already contained commas.

This commit adds methods for escaping commas already in tags ("," gets
replaced with "\,") so that the original tags can be correctly restored
from the text that gets stored in the db.