First pass at implementing all of Text
This is my first pass at implementing all of text. Image would follow a  similar pattern. This compiles and runs (without the unimplemented interface adaptors which I do not believe are my job), but doesn't display anything. I was largely trying to follow the form of Dexter's work with Pen but I am missing something. It is very very late now so I am going to go to bed and will probably have to finish this on Monday. @dexgs  can you take a look at why this isn't working? I think it's to do with how I am setting things up in GUITextBox. Gonna sleep now!