Add ResourceLoader utility class
The project currently stores fonts, images, etc. in the `res` directory
and just loads them by path, but this won't work if we want to build the
project into a JAR file so we can just give it to someone to run.

I modified build.gradle to add the `res` directory to the JAR file and
added the ResourceLoader utility class to provide methods which load
resources either from the external `res` directory or from inside the
current JAR file, depending on what is available.

I changed loading the stylesheets in `gui.Main` from using a hard-coded
path to using the ResourceLoader class.

Additionally, I added helper methods for loading icon images in the
ResourceLoader and used them to add icons to the pen tool's toolbar
button and to the tool settings pane.