Modify Page to support MediaTool
Add a new method, `setUIlayer`, which allows setting a Node to be
displayed above the contents of the page. This allows showing UI
controls "on" the page which are not themselves page contents.

In the case of the MediaTool, the `setUIlayer` allows displaying a box
selection rectangle over the page contents when the user clicks and
drags to select multiple media in the page.

---

Add a new method, `getAllMedia`, which returns all the Media with
assigned IDs which are currently in the page.

---

Add a new method, `getCoordsInv` which transforms coordinates out of the
page's coordinate space.