Improve ColourTool (Again)
Instead of checking for the sliders being "pressed" which only evaluates
to true when they are manipulated with the mouse, check that they are
focused, which allows the colour settings GUI to update properly when
they keyboard is used to manipulate the sliders.

Additionally, add increments to the sliders so that moving them with the
keyboard changes the value by an appropriate amount.