Add MediaTool
Add a tool which allows changing the following properties common to all
Media types:

- Name
- Position
- Rotation
- Z-index

The tool also allows the deletion of Media.

The actual modification of these fields is straightforward, but the user
interaction more involved.

The media tool allows selecting multiple media at once. This can be done
either with a click-and-drag box selection, or by picking media one by
one by clicking.

A visual representation of the box selection is provided by the new
`BoxSelectRectangle` class.

Also, when a selection is made, the settings gui for the media tool
tries to adjust to the current selection. The tool modifies the selected
media all at once, but the selected media may not all have the same
values for the properties modified by this tool. Therefore, the tool has
to pick appropriate starting values for the settings based on the
current selection. The implemented strategy is the following:

- The name field is set to the same of the selected media if they all
  have the same name and is left empty otherwise
- The angle slider is set to the average angle of rotation of the
  selection
- The Z-Index spinner is set to the maximum Z-Index of the selection.

The selected media can be moved by clicking and dragging any currently
selected media.