Make MediaCommunicator generate IDs
Instead of having to call `getNewID()` on MediaCommunicator before
calling `updateMedia`, make the MediaCommunicator generate the ID
itself.

If the caller wants to do something with the ID before any
MediaObservers are notified, it can supply a `Consumer<Long>` to the
`updateMedia` method which will be called AFTER the ID is generated, but
BEFORE any MediaObservers are notified of the change.