Fix: SQLiteStorage close PreparedStatements and use File URL
Make sure all PreparedStatements are properly obtained in
try-with-resources blocks so that they get closed properly.

When saving a file, use file.toURI().toURL() instead of file.toPath().
This fixes saving to files whose paths contain whitespace and other
problematic characters.

Finally, added a method to load a Media object from just the properties
defined in the Media class. This avoids the problem of